refactor: 애그리게이트 내부를 레이어 기준으로 분리
