+++ src/main/java/com/techfork/activity/bookmark/infrastructure/BookmarkRepository.java
@@ -1,7 +1,7 @@
-package com.techfork.activity.bookmark.repository;
+package com.techfork.activity.bookmark.infrastructure;
 
-import com.techfork.activity.bookmark.dto.BookmarkDto;
-import com.techfork.activity.bookmark.entity.Bookmark;
+import com.techfork.activity.bookmark.application.BookmarkDto;
+import com.techfork.activity.bookmark.domain.Bookmark;
 import com.techfork.domain.post.entity.Post;
 import com.techfork.domain.useraccount.entity.User;
 import org.springframework.data.domain.Pageable;
@@ -15,7 +15,7 @@ import java.util.Optional;
 public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
 
     @Query("""
-            SELECT new com.techfork.activity.bookmark.dto.BookmarkDto(
+            SELECT new com.techfork.activity.bookmark.application.BookmarkDto(
                 b.id, p.id, p.title, p.shortSummary, p.url, t.companyName, t.logoUrl,
                             p.publishedAt, p.thumbnailUrl, p.viewCount, null, true
             )
