+++ src/test/java/com/techfork/activity/bookmark/infrastructure/BookmarkRepositoryTest.java
@@ -1,7 +1,7 @@
-package com.techfork.activity.bookmark.repository;
+package com.techfork.activity.bookmark.infrastructure;
 
-import com.techfork.activity.bookmark.dto.BookmarkDto;
-import com.techfork.activity.bookmark.entity.Bookmark;
+import com.techfork.activity.bookmark.application.BookmarkDto;
+import com.techfork.activity.bookmark.domain.Bookmark;
 import com.techfork.domain.post.entity.Post;
 import com.techfork.domain.post.repository.PostRepository;
 import com.techfork.domain.source.entity.TechBlog;
