+++ src/test/java/com/techfork/activity/bookmark/presentation/BookmarkIntegrationTest.java
@@ -1,9 +1,8 @@
-package com.techfork.activity.bookmark.integration;
+package com.techfork.activity.bookmark.presentation;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.techfork.activity.bookmark.dto.BookmarkRequest;
-import com.techfork.activity.bookmark.entity.Bookmark;
-import com.techfork.activity.bookmark.repository.BookmarkRepository;
+import com.techfork.activity.bookmark.domain.Bookmark;
+import com.techfork.activity.bookmark.infrastructure.BookmarkRepository;
 import com.techfork.domain.post.entity.Post;
 import com.techfork.domain.post.repository.PostRepository;
 import com.techfork.domain.source.entity.TechBlog;
