+++ src/test/java/com/techfork/domain/useraccount/repository/UserRepositoryTest.java
@@ -1,10 +1,10 @@
 package com.techfork.domain.useraccount.repository;
 
 import com.techfork.activity.readpost.entity.ReadPost;
-import com.techfork.activity.bookmark.entity.Bookmark;
+import com.techfork.activity.bookmark.domain.Bookmark;
 import com.techfork.activity.readhistory.entity.SearchHistory;
 import com.techfork.activity.readpost.repository.ReadPostRepository;
-import com.techfork.activity.bookmark.repository.BookmarkRepository;
+import com.techfork.activity.bookmark.infrastructure.BookmarkRepository;
 import com.techfork.activity.readhistory.repository.SearchHistoryRepository;
 import com.techfork.domain.post.entity.Post;
 import com.techfork.domain.post.repository.PostRepository;
